# app/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/composite/recipes/id/{recipe_id}": {"get": {"summary": "Get Recipe", "operationId": "get_recipe_composite_recipes_id__recipe_id__get", "parameters": [{"name": "recipe_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Recipe Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/composite/nutrition/{nutrition_id}": {"get": {"summary": "Get Nutrition", "operationId": "get_nutrition_composite_nutrition__nutrition_id__get", "parameters": [{"name": "nutrition_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Nutrition Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/composite/mealplan/{mealplan_id}": {"get": {"summary": "Get Mealplan", "operationId": "get_mealplan_composite_mealplan__mealplan_id__get", "parameters": [{"name": "mealplan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Mealplan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Mealplan": {"properties": {"meal_id": {"type": "integer", "title": "Meal Id"}, "breakfast_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Breakfast Recipe"}, "lunch_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Lunch Recipe"}, "dinner_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Dinner Recipe"}, "links": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Links"}}, "type": "object", "required": ["meal_id"], "title": "Mealplan", "example": {"breakfast_recipe": 171, "dinner_recipe": 192, "links": {"breakfast_recipe": "/recipes/171", "delete": "/mealplans/10/delete", "dinner_recipe": "/recipes/192", "edit": "/mealplans/10/edit", "lunch_recipe": "/recipes/180", "self": "/mealplans/10"}, "lunch_recipe": 180, "meal_id": 10}}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}